fix: verify session module rename
